03_hello_agent/hello_sync_agent-app/pyproject.toml
[project]
name = "hello-sync-agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
